Better pycode/snakecode parsing (2)
Moved `get_next_queriable` function to parser.py.
Will now update `block_indent` variable so that it's always correct.
